Create new Stripe handler and refactor functions
This commit involves the creation of a new Stripe handler module, separate from the existing helper module. The handling of webhooks, initially done in the helper, is now shifted to this new handler module. The change also includes the necessary adjustments in http.js to point to the new StripeHandler module.