Add support for multiple pricing plans
The `createCheckout` function has been updated in `helper.js` to support multiple pricing plans. This allows users to pass a new optional `plan` parameter, which will default to the first one if not provided. The `websocket.js` and `http.js` files also pass this `plan` parameter when calling the `createCheckout` function. Updates were made to the `README.md` to reflect these changes.